pyproject.toml
[project]
name = "ytm"
version = "0.1.0"
description = "YouTube Playlist Manager with time-travel capabilities"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "duckdb>=1.0.0",
    "yt-dlp>=2024.0.0",
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "pytz>=2024.0",
]

[project.scripts]
ytm = "ytm.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ytm"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
